requirements.txt
annotated-types==0.7.0
anyio==4.4.0
boto3==1.35.3
botocore==1.35.3
click==8.1.7
dynamodb-json==1.4.2
fastapi==0.112.1
h11==0.14.0
idna==3.7
iniconfig==2.0.0
jmespath==1.0.1
packaging==24.1
pluggy==1.5.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
